requirements.txt
openai==0.27.8
python-dotenv==1.0.1
pyttsx3==2.90
SpeechRecognition==3.10.0
